+++ src/components/Buttons/BaseButton/BaseButton.component.tsx
@@ -1,12 +1,12 @@
 import React, { FC } from 'react';
 import styled from 'styled-components';
 
-import { ButtonProps } from './Button.types';
+import { ButtonProps } from '../common';
 
 const BaseButtonStyledComponent = styled.button<ButtonProps>(
     ({ styles }) => `${styles}`
 );
 
-export const ButtonBase: FC<ButtonProps> = (props) => {
+export const BaseButton: FC<ButtonProps> = (props) => {
     return <BaseButtonStyledComponent {...props} />;
 };
