+++ src/components/Buttons/BaseButton/BaseButton.test.tsx
@@ -2,11 +2,11 @@ import * as React from 'react';
 import '@testing-library/jest-dom';
 import { render, screen } from '@testing-library/react';
 
-import { ButtonBase } from './ButtonBase.component';
+import { BaseButton } from './BaseButton.component';
 
 describe('Running tests for ButtonBase', () => {
     test('Given disabled prop then button disabled', () => {
-        render(<ButtonBase children={'ButtonBase'} disabled />);
+        render(<BaseButton children={'ButtonBase'} disabled />);
         expect(
             screen.getByRole('button', { name: 'ButtonBase' })
         ).toBeDisabled();
@@ -14,7 +14,7 @@ describe('Running tests for ButtonBase', () => {
 
     test('Given styles passed explicitly, when rendering button, button takes on dynamically passed styles ', () => {
         const expectedCss = 'padding: 20px 20px';
-        render(<ButtonBase children={'ButtonBase'} styles={expectedCss} />);
+        render(<BaseButton children={'ButtonBase'} styles={expectedCss} />);
         expect(screen.getByRole('button', { name: 'ButtonBase' })).toHaveStyle(
             expectedCss
         );
