--- src/components/Buttons/base/index.ts
@@ -1,4 +1,0 @@
-export * from './Button.constants';
-export * from './Button.types';
-export * from './ButtonBase.component';
-export * from './TextButtonBase.component';
