+++ src/components/Buttons/common/types/index.ts
@@ -1,0 +1,1 @@
+export * from './Button.types';
