+++ src/components/Buttons/index.ts
@@ -1,4 +1,2 @@
-export * from './AnimatedButton.component';
-export * from './PlainButton.component';
-export * from './base';
-export * from './utils';
+export * from './AnimatedButton/AnimatedButton.component';
+export * from './PlainButton/PlainButton.component';
