+++ src/core/constants/index.ts
@@ -1,4 +1,2 @@
 export * from './Colors.constants';
-export * from './Colors.types';
-export * from './Sizes.types';
 export * from './DefaultSizes.constants';
