+++ src/core/types/index.ts
@@ -1,0 +1,2 @@
+export * from './Colors.types';
+export * from './Sizes.types';
