--- src/index.ts
@@ -1,2 +1,0 @@
-export * from './components/Buttons';
-export * from './config';
